stages 1 to 7 in main works aka not getting stuck and does what suppose to do.
only thing might not be ok is the use of nFeatures in the descriptors function (im asking friends about it..)

all other stages work fine and fluent including all the allocationgs on matrix

the functions i wrote in sp_image.. also suppose to work fine since the openCV parts are the same as moab did. the functions calc distances suppose to work also since they are easy.